fix(dev): silence edge-runtime warnings from the instrumentation hook
